--- client/src/components/Sidebar/Sidebar.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Sidebar() {
-  return (
-    <div>Sidebar</div>
-  )
-}
